requirements.txt
streamlit
langchain
langchain-community
openai
tiktoken
python-dotenv
faiss-cpu
requests
pydantic
beautifulsoup4